% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_initialization_functs.R
\name{append_noise_to_weights}
\alias{append_noise_to_weights}
\title{Append noise to weights}
\usage{
append_noise_to_weights(w, n_rep, sd)
}
\arguments{
\item{w}{initial weight vector}

\item{n_rep}{number of noisy columns to append to w}

\item{sd}{standard deviation of noise}
}
\value{
initial weight matrix
}
\description{
Adds Gaussian noise to an initial weight vector.
}
