% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_object_functs.R
\name{augment_family_object}
\alias{augment_family_object}
\title{Augment family object}
\usage{
augment_family_object(f)
}
\arguments{
\item{f}{a family object}
}
\value{
the augmented family object
}
\description{
Augments a family object, adding the fields "skewness," "mu.eta.prime, "simulate_from_mus," and "log_py_given_mu."
}
\examples{
pois_augmented <- augment_family_object(poisson())
}
