% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_col_names}
\alias{check_col_names}
\title{Check column names}
\usage{
check_col_names(covariate_matrix)
}
\arguments{
\item{covariate_matrix}{a matrix of observed covariates}
}
\description{
Checks if "perturbation" is a column name of covariate_matrix; if so, produces an error.
}
