% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_matrix_functs.R
\name{compute_info_submat}
\alias{compute_info_submat}
\alias{compute_info_submat_1}
\alias{compute_info_submat_23}
\alias{compute_info_submat_4}
\alias{compute_info_submat_56}
\title{Compute information submatrix}
\usage{
compute_info_submat_1(pieces)

compute_info_submat_23(pieces, k)

compute_info_submat_4(pieces)

compute_info_submat_56(pieces, k)
}
\arguments{
\item{pieces}{"pieces list," as outputted by get_info_mat_pieces.}

\item{k}{either "m" or "g;" required for compute_info_submat_23 and compute_info_submat_56.}
}
\value{
the requested information submatrix
}
\description{
Compute information submatrices 1-6. For matrices 2-3 and 5-6, specify either "m" or "g."
}
