% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{compute_theoretical_conditional_means}
\alias{compute_theoretical_conditional_means}
\title{Compute theoretical conditional means}
\usage{
compute_theoretical_conditional_means(
  intercept,
  perturbation_coef,
  fam,
  covariate_matrix = NULL,
  covariate_coefs = NULL,
  offset = NULL
)
}
\arguments{
\item{intercept}{intercept term}

\item{perturbation_coef}{coefficient corresponding to perturbation}

\item{fam}{family object describing response distribution}

\item{covariate_matrix}{(optional) matrix of technical covariates}

\item{covariate_coefs}{(optional) coefficients corresponding to technical covariates}

\item{offset}{(optional) vector of offsets}
}
\value{
the scalar (or vector, if covariate_matrix is supplied) of conditional means
}
\description{
Computes the conditional means of the GLM-EIV model given the ground truth.
}
