% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{diag_matrix_multiply}
\alias{diag_matrix_multiply}
\title{Diagonal matrix multiply}
\usage{
diag_matrix_multiply(A, D, B)
}
\arguments{
\item{A}{a matrix}

\item{D}{a diagonal matrix, represented as a numeric vector}

\item{B}{a matrix}
}
\value{
the matrix product ADB.
}
\description{
Performs the matrix multiplication ADB, where D is diagonal (and A and B are
not necessarily diagonal).
}
\examples{
\dontrun{
B <- matrix(sample(0:1, size = 12, TRUE), ncol = 3)
D <- c(2, 3, 4, 5)
A <- t(B)
all.equal(A \%*\% diag(D) \%*\% B, diag_matrix_multiply(A, D, B))
}
}
