% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_initialization_functs.R
\name{flip_weights}
\alias{flip_weights}
\title{Flip weights}
\usage{
flip_weights(w, p)
}
\arguments{
\item{w}{a binary (0/1) vector}

\item{p}{the expected fraction of weights to flip}
}
\value{
a new binary vector with E(p) weights flipped.
}
\description{
Flip weights
}
\examples{
w <- rbinom(n = 100, size = 1, prob = 0.5)
p <- 0.1
glmeiv:::flip_weights(w, p)
}
