% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{generate_glm_data_sim}
\alias{generate_glm_data_sim}
\title{Generate glm data for simulation}
\usage{
generate_glm_data_sim(
  intercept,
  perturbation_coef,
  perturbation_indicators,
  fam,
  covariate_matrix,
  covariate_coefs,
  offset,
  n,
  B
)
}
\arguments{
\item{intercept}{intercept term}

\item{perturbation_coef}{coefficient for perturbation}

\item{perturbation_indicators}{an n times B matrix of perturbation indicators}

\item{fam}{family object describing response}

\item{covariate_matrix}{fixed matrix of covariates}

\item{covariate_coefs}{coefficients for technical factors}

\item{offset}{optional offset vector}

\item{n}{the number of examples}

\item{B}{the number of datasets to resample}
}
\value{
an n times B matrix of synthetic response data
}
\description{
A lower-level function to generate GLM-EIV data.
}
