% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_initialization_functs.R
\name{get_bayes_optimal_bdy}
\alias{get_bayes_optimal_bdy}
\title{Get Bayes-optimal boundary}
\usage{
get_bayes_optimal_bdy(l_0, l_1, pi, fam)
}
\arguments{
\item{l_0}{l_i(0)}

\item{l_1}{l_i(1)}

\item{pi}{probability of perturbation}

\item{fam}{augmented family object}
}
\value{
the Bayes-optimal decision boundary
}
\description{
Given an augmented family object, linear component corresponding to P = 0,
linear component corresponding to P = 1, and pi, get the Bayes-optimal
decision boundary.
}
\examples{
fam <- poisson() \%>\% augment_family_object()
l_0 <- -1
l_1 <- 2
pi <- 0.2
get_bayes_optimal_bdy(l_0, l_1, pi, fam)
}
