% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_matrix_functs.R
\name{get_info_mat_pieces_helper}
\alias{get_info_mat_pieces_helper}
\title{Get matrices from fit}
\usage{
get_info_mat_pieces_helper(fit, idxs)
}
\arguments{
\item{fit}{a fitted GLM object}

\item{idxs}{the sample indexes on which to extract the matrices}
}
\value{
a list of diagonal matrices, represented as list of numeric vectors. The diagonal matrices are
Delta, V, Delta', H, and Delta * H.
}
\description{
Obtains the matrices Delta, V, Delta', and H, and Delta * H from a fitted GLM object
given a set of indexes.
}
