% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_em_algo_functs.R
\name{get_marginal_mixture_weights}
\alias{get_marginal_mixture_weights}
\title{Get marginal mixture weights}
\usage{
get_marginal_mixture_weights(v, fam, offset)
}
\arguments{
\item{v}{a vector (of mRNA or gRNA counts)}

\item{fam}{family object describing the counts}

\item{offset}{optional vector of linear offsets}
}
\value{
the EM weights
}
\description{
Fit a marginal mixture model; get the (correctly oriented) weights
}
