% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_initialization_functs.R
\name{get_optimal_threshold}
\alias{get_optimal_threshold}
\title{Get optimal threshold}
\usage{
get_optimal_threshold(
  g_intercept,
  g_perturbation,
  g_fam,
  pi,
  covariate_matrix = NULL,
  g_covariate_coefs = NULL,
  g_offset = NULL
)
}
\arguments{
\item{g_intercept}{intercept for gRNA model}

\item{g_perturbation}{perturbation coefficient for gRNA model}

\item{g_fam}{family object describing distribution of g}

\item{pi}{probability of perturbation}

\item{covariate_matrix}{(optional) matrix of technical covariates}

\item{g_covariate_coefs}{(optional) coefficients corresponding to technical factors}

\item{g_offset}{(optional) the offset vector}
}
\value{
an integer (0/1) vector of thresholded counts.
}
\description{
Thresholds counts according to the Bayes-optimal decision boundary. When a covariate matrix is present, the decision boundary that is used is the mean of the Bayes-optimal decision boundaries across all examples.
}
\examples{
library(magrittr)
m_fam <- g_fam <- poisson() \%>\% augment_family_object()
m_intercept <- 2; m_perturbation <- -1; g_intercept <- -2; g_perturbation <- 1
pi <- 0.2; n <- 1000; B <- 500; alpha <- 0.95
m_offset <- g_offset <- NULL
bdy <- get_optimal_threshold(g_intercept, g_perturbation, g_fam, pi)
}
