% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_quick_simulated_data}
\alias{get_quick_simulated_data}
\title{Get quick simulated data}
\usage{
get_quick_simulated_data(n = 1000)
}
\arguments{
\item{n}{number of examples}
}
\value{
a list containing the entries m, g, p, pi, covariate_matrix, m_coef, and g_coef.
}
\description{
Quickly generates simulated data.
}
