% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory_functs.R
\name{get_thresholding_estimator_est_no_int}
\alias{get_thresholding_estimator_est_no_int}
\title{Get thresholding estimator estimate (no intercept model)}
\usage{
get_thresholding_estimator_est_no_int(c, g_beta, pi, m_beta)
}
\arguments{
\item{c}{selected threshold}

\item{g_beta}{perturbation coefficient for gRNA model}

\item{pi}{probability of perturbation}

\item{m_beta}{perturbation coefficient for mRNA model}
}
\value{
limit (in probability) of estimator
}
\description{
Get thresholding estimator estimate (no intercept model)
}
