% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory_functs.R
\name{get_thresholding_estimator_var_no_int}
\alias{get_thresholding_estimator_var_no_int}
\title{Get variance of thresholding estimator (no intercept)}
\usage{
get_thresholding_estimator_var_no_int(c, g_beta, pi, m_beta, n)
}
\arguments{
\item{c}{selected threshold}

\item{g_beta}{perturbation coefficient for gRNA model}

\item{pi}{probability of perturbation}

\item{m_beta}{perturbation coefficient for mRNA model}

\item{n}{number of cells}
}
\value{
theoretical variance of estimator
}
\description{
Returns variance of thresholding estimator in no-intercept model.
}
