% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functs.R
\name{glm_log_lik}
\alias{glm_log_lik}
\title{GLM log lik}
\usage{
glm_log_lik(object)
}
\arguments{
\item{object}{a fitted GLM}
}
\value{
its log-likelihood
}
\description{
Internal function; return the log-likelihood of a GLM object.
}
