% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmeiv.R
\docType{package}
\name{glmeiv}
\alias{glmeiv}
\title{glmeiv: a glm-oriented variant of the errors-in-variables model.}
\description{
GLM-EIV is GLM-oriented variant of the errors-in-variables model. GLM-EIV models the response as
a GLM of an unobserved predictor. A noisy version of the predictor is obeserved instead, which
itself is modeled as a GLM of the true predictor. glmeiv implements estimation and inference in the GLM-EIV model.
}
