% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_initialization_functs.R
\name{initialize_weights_using_marginal_mixtures}
\alias{initialize_weights_using_marginal_mixtures}
\title{Initialize weights using marginal mixtures}
\usage{
initialize_weights_using_marginal_mixtures(
  m,
  g,
  m_fam,
  g_fam,
  m_offset,
  g_offset,
  lambda = NULL
)
}
\arguments{
\item{m}{mRNA counts}

\item{g}{gRNA counts}

\item{m_fam}{family describing mRNA counts}

\item{g_fam}{family describing gRNA counts}

\item{m_offset}{optional offsets for m}

\item{g_offset}{optional offsets for g}

\item{lambda}{optional weight in weighted average; if not supplied, chosen adaptively}
}
\value{
initial weights for algorithm
}
\description{
Initializes weights for GLM-EIV by fitting marginal mixtures, then taking weighted average.
}
