% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functs.R
\name{lm_log_lik}
\alias{lm_log_lik}
\title{LM log lik}
\usage{
lm_log_lik(object)
}
\arguments{
\item{object}{a linear model (or GLM with Gaussian family, identity link)}
}
\value{
its log-likelihood
}
\description{
Internal function; return the log-likelihood of an lm.
}
