% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{obtain_valid_ids}
\alias{obtain_valid_ids}
\title{Obtain valid IDs}
\usage{
obtain_valid_ids(
  sim_res,
  spread_thresh = 0.1,
  approx_0_thresh = 50,
  approx_1_thresh = 50,
  g_pert_lower = -0.5,
  g_pert_upper = Inf,
  m_pert_lower = -Inf,
  m_pert_upper = 0.5,
  pi_lower = 0,
  pi_upper = 0.5
)
}
\arguments{
\item{sim_res}{sim_res object}

\item{spread_thresh}{spread threshold}

\item{approx_0_thresh}{number cells approximately 0 threshold}

\item{approx_1_thresh}{number cells approximately 1 threshold}

\item{g_pert_lower}{minimum value for g_pert}

\item{g_pert_upper}{maximum value for g_pert}

\item{m_pert_lower}{minimum value for m_pert}

\item{m_pert_upper}{maximum value for m_pert}

\item{pi_lower}{minimum value for pi}

\item{pi_upper}{maximum value for pi}
}
\value{
a list of length two; (i) a data frame giving the classification of each EM run, and (ii) a character vector of valid IDs for both EM and thresholding
}
\description{
Obtains valid IDs given the output of a simulatr experiment.
}
