% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functs.R
\name{plot_count_distribution}
\alias{plot_count_distribution}
\title{Plot count distribution}
\usage{
plot_count_distribution(generated_data, modality)
}
\arguments{
\item{generated_data}{a list containing p, m, and/r g.}

\item{modality}{either "mRNA" or "gRNA"}
}
\value{
a ggplot of the histogram
}
\description{
Plots the distribution of m or g counts, colored by perturbation status.
}
\examples{
\dontrun{
n <- 10000
m_fam <- g_fam <- poisson()
pi <- 0.4
covariate_matrix <- NULL
m_coef <- c(1, -2)
g_coef <- c(-2, 3)
generated_data <- generate_data_from_model(m_fam, g_fam,
m_coef, g_coef, pi, covariate_matrix, n = n)
p <- plot_count_distribution(generated_data, "gRNA")
plot(p)
}
}
