% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functs.R
\name{plot_em_classifications}
\alias{plot_em_classifications}
\title{Plot EM classifications}
\usage{
plot_em_classifications(
  em_classifications,
  sim_spec,
  grid_row_id,
  categories = c("confident-plausible", "confident-implausible",
    "unconfident-plausible", "unconfident-implausible"),
  parameter = "m_perturbation"
)
}
\arguments{
\item{em_classifications}{dataframe outputted by \code{obtain_valid_ids}}

\item{sim_spec}{a simulatr specifier object (to supply the ground truth)}

\item{grid_row_id}{grid_row_id giving row to plot}
}
\value{
a histogram (in ggplot format)
}
\description{
Creates a histogram containing the EM estimates for m_perturbation colored by type.
}
\details{
There are four categories: the cartesian product of (confident vs. unconfident) and (pluasible vs. implausible).
An estimate is "confident" if the posterior membership probabilities are sufficiently spread out (see function
obtain valid IDs); and estimate is "plausible" if the estimates themselves are in biologically realistic ranges.
Valid IDs are those that are both confident and plausible.
}
