% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functs.R
\name{plot_mixture}
\alias{plot_mixture}
\title{Plot mixture}
\usage{
plot_mixture(
  density_df,
  thresh = NULL,
  x_max = Inf,
  x_min = -Inf,
  points = TRUE,
  xlab = NULL
)
}
\arguments{
\item{density_df}{a density df as outputted by "get_theoretical_densities_and_thresholds."}

\item{thresh}{a threshold to plot as a vertical line}

\item{x_max}{maximal x-value}

\item{x_min}{minimal x-value}

\item{points}{plot points? (if FALSE, only lines)}

\item{xlab}{x-axis label}
}
\value{
a ggplot object
}
\description{
Plots a mixture distribution.
}
