% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functs.R
\name{plot_posterior_membership_probabilities}
\alias{plot_posterior_membership_probabilities}
\title{Plot posterior membership probabilities}
\usage{
plot_posterior_membership_probabilities(
  sim_res,
  grid_row_id,
  valid_ids,
  n_approx_01 = 50
)
}
\arguments{
\item{sim_res}{raw result data frame outputted by simulatr}

\item{grid_row_id}{an grid row id}

\item{valid_ids}{character vector of valid IDs}
}
\value{
a list of length 2: (i) a list of histogram plots, and (ii) a data frame storing summary metrics for each run
}
\description{
Plots all sets of posterior membership probabilities for a given grid_row_id of a sim_res data frame.
}
