% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{process_glmeiv_results_simulatr}
\alias{process_glmeiv_results_simulatr}
\title{Process GLM-EIV results simulatr}
\usage{
process_glmeiv_results_simulatr(
  em_fit,
  s,
  dat,
  save_membership_probs_mult,
  time
)
}
\arguments{
\item{em_fit}{the fitted GLM-EIV model}

\item{s}{data frame of fitted coefficients and inferences}

\item{dat}{the data}

\item{save_membership_probs_mult}{integer giving multiple to save membership probabilities}

\item{time}{execution time (in s)}
}
\value{
a processed, long data frame
}
\description{
Processes the results outputted by a GLM-EIV method for use in simulatr.
}
