% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_initialization_functs.R
\name{random_initialization}
\alias{random_initialization}
\title{Random initialization}
\usage{
random_initialization(n, pi)
}
\arguments{
\item{n}{length of output vector}

\item{pi}{fraction of entries to set to 1}
}
\value{
randomly initialized vector
}
\description{
Output a vector of length n; n * pi of the entries are randomly set to 1, all others to 0.
}
