% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_matrix_functs.R
\name{run_delta_method}
\alias{run_delta_method}
\title{Run delta method}
\usage{
run_delta_method(estimate, std_error, alpha = 0.95)
}
\arguments{
\item{estimate}{an estimate}

\item{std_error}{the standard error of the estimate}

\item{alpha}{confidence level; defauly 0.95}
}
\value{
a data frame with columns estimate, confint_lower, confint_upper
}
\description{
Computes the estimate and CI exp(estimate), given its standard error.
}
