% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_em_algo_functs.R
\name{run_em_algo_mixture_init}
\alias{run_em_algo_mixture_init}
\title{Run EM algorithm -- mixture model initialization}
\usage{
run_em_algo_mixture_init(
  dat,
  g_fam,
  m_fam,
  covariate_matrix,
  m_offset,
  g_offset,
  alpha,
  n_em_rep,
  sd = 0.15,
  save_membership_probs_mult = 250,
  lambda = NULL
)
}
\arguments{
\item{dat}{the count data; should have columns m and g.}

\item{g_fam}{family object used to model gRNA counts.}

\item{m_fam}{family object used to model mRNA counts.}

\item{covariate_matrix}{optional matrix of covariates}

\item{m_offset}{optional offsets for mRNA model}

\item{g_offset}{optional offsets for gRNA model}

\item{alpha}{returns a (1-alpha)\% confidence interval}

\item{n_em_rep}{number of replicates of em algorithm}

\item{sd}{sd of noise to add to initial weights}

\item{save_membership_probs_mult}{save the posterior membership probabilities at this multiple}

\item{lambda}{mixing parameter for weighted average of weights; default NULL chooses mixing parameter adaptively.}
}
\value{
a tibble with columns parameter, target (fields estimate, std_error, p-value, confint lower, and confint higher), and value.
}
\description{
Run EM algorithm -- mixture model initialization
}
\examples{
\dontrun{
library(magrittr)
m_fam <- g_fam <- poisson() \%>\% augment_family_object()
m_intercept <- 2; m_perturbation <- -1; g_intercept <- -1; g_perturbation <- 2
pi <- 0.2; n <- 2000; B <- 5; alpha <- 0.95; n_em_rep <- 3;
sd <- 0.15; lambda <- NULL
m_offset <- g_offset <- NULL
m_covariate_coefs <- g_covariate_coefs <- covariate_matrix <- NULL
dat_list <- generate_full_data(m_fam, m_intercept, m_perturbation, g_fam,
g_intercept, g_perturbation, pi, n, B, covariate_matrix,
m_covariate_coefs, g_covariate_coefs, m_offset, g_offset)
dat <- dat_list[[1]]
run_em_algo_mixture_init(dat, g_fam, m_fam, covariate_matrix,
m_offset, g_offset, alpha, n_em_rep, sd)
}
}
