% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_em_algo_functs.R
\name{run_em_algo_simulatr_optimal_thresh}
\alias{run_em_algo_simulatr_optimal_thresh}
\title{Run em algorithm for simulatr using optimal threshold initialization}
\usage{
run_em_algo_simulatr_optimal_thresh(
  dat_list,
  g_intercept,
  g_perturbation,
  g_fam,
  m_fam,
  pi,
  covariate_matrix,
  g_covariate_coefs,
  m_offset,
  g_offset,
  alpha,
  n_em_rep,
  p_flip
)
}
\arguments{
\item{dat_list}{a list of data frames, each of which has columns m, g, p.}

\item{g_intercept}{the intercept for the gRNA model.}

\item{g_perturbation}{the perturbation coefficient for the gRNA model.}

\item{g_fam}{family object describing gRNA model.}

\item{m_fam}{family object describing mRNA model.}

\item{pi}{probability of perturbation}

\item{covariate_matrix}{data frame of technical factors; can be null}

\item{g_covariate_coefs}{technical factor coefficients for gRNA model}

\item{m_offset}{optional offset vector for mRNA model}

\item{g_offset}{optional offset vector for gRNA model}

\item{alpha}{(1-alpha)\% CIs returned}

\item{n_em_rep}{number of EM algorithm runs to conduct}

\item{p_flip}{probability of flipping a given initialization weight.}
}
\value{
a data frame with columns parameter, target, value, and run_id
}
\description{
Run em algorithm for simulatr using optimal threshold initialization
}
\examples{
library(magrittr)
m_fam <- g_fam <- poisson() \%>\% augment_family_object()
m_intercept <- 2; m_perturbation <- -1; g_intercept <- -2; g_perturbation <- 1
pi <- 0.2; n <- 1000; B <- 500; alpha <- 0.95; n_em_rep <- 5; p_flip <- 0.01
m_offset <- g_offset <- NULL
m_covariate_coefs <- g_covariate_coefs <- covariate_matrix <- NULL
dat_list <- generate_full_data(m_fam, m_intercept, m_perturbation, g_fam,
g_intercept, g_perturbation, pi, n, B, covariate_matrix,
m_covariate_coefs, g_covariate_coefs, m_offset, g_offset)
run_em_algo_simulatr_optimal_thresh(dat_list, g_intercept, g_perturbation,
g_fam, m_fam, pi, covariate_matrix, g_covariate_coefs, m_offset, g_offset,
alpha, n_em_rep, p_flip)
}
