% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_em_algo_functs.R
\name{run_full_glmeiv_given_fitted_means}
\alias{run_full_glmeiv_given_fitted_means}
\title{Run full GLM-EIV given fitted means}
\usage{
run_full_glmeiv_given_fitted_means(
  m_fam,
  g_fam,
  m,
  g,
  m_mus_pert0,
  m_mus_pert1,
  g_mus_pert0,
  g_mus_pert1,
  fit_pi,
  covariate_matrix,
  m_offset,
  g_offset,
  ep_tol = 1e-04,
  max_it = 75
)
}
\arguments{
\item{m_fam}{family describing m}

\item{g_fam}{family describing g}

\item{m}{mRNA counts}

\item{g}{gRNA counts}

\item{m_mus_pert0}{fitted (or hypothesized) means mu^m(0)}

\item{m_mus_pert1}{fitted (or hypothesized) means mu^m(1)}

\item{g_mus_pert0}{fitted (or hypothesized) means mu^g(0)}

\item{g_mus_pert1}{fitted (or hypothesized) means mu^g(1)}

\item{fit_pi}{fitted (or hypothesized) value for pi}

\item{covariate_matrix}{the matrix of covariates; NULL if there are no covariates.}

\item{m_offset}{offsets for m}

\item{g_offset}{offsets for g}

\item{ep_tol}{(optional) EM convergence threshold}

\item{max_it}{(optional) maximum number of EM iterations}
}
\value{

}
\description{
Run full GLM-EIV given fitted means
}
