% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_em_algo_functs.R
\name{run_full_glmeiv_given_pilot_params}
\alias{run_full_glmeiv_given_pilot_params}
\title{Run full GLM-EIV given pilot estiamtes}
\usage{
run_full_glmeiv_given_pilot_params(
  m,
  g,
  m_fam,
  g_fam,
  pi_guess,
  m_intercept_guess,
  m_perturbation_guess,
  m_covariate_coefs_guess,
  g_intercept_guess,
  g_perturbation_guess,
  g_covariate_coefs_guess,
  covariate_matrix,
  m_offset,
  g_offset,
  ep_tol = 1e-05,
  max_it = 75
)
}
\arguments{
\item{m}{mRNA counts}

\item{g}{gRNA counts}

\item{m_fam}{family describing m}

\item{g_fam}{family describing g}

\item{pi_guess}{pilot guess for pi}

\item{m_intercept_guess}{pilot guess for m_intercept}

\item{m_perturbation_guess}{pilot guess for m_perturbation}

\item{m_covariate_coefs_guess}{pilot guess for m_covariate_coefficients}

\item{g_intercept_guess}{pilot guess for g_intercept}

\item{g_perturbation_guess}{pilot guess for g_perturbation}

\item{g_covariate_coefs_guess}{pilot guess for g_covariate_coefs}

\item{covariate_matrix}{the matrix of covariates; NULL if there are no covariates.}

\item{m_offset}{offsets for m}

\item{g_offset}{offsets for g}

\item{ep_tol}{(optional) EM convergence threshold}

\item{max_it}{(optional) maximum number of EM iterations}
}
\value{
the fitted GLM-EIV object.
}
\description{
Run full GLM-EIV given pilot estiamtes
}
\examples{
set.seed(4)
m_fam <- g_fam <- augment_family_object(poisson())
n <- 5000
lib_size <- rpois(n = n, lambda = 5000)
m_offsets <- g_offsets <- log(lib_size)
pi <- 0.3
m_intercept <- log(0.05)
m_perturbation <- log(0.75)
g_intercept <- log(0.025)
g_perturbation <- log(1.4)
covariate_matrix <- data.frame(batch = rbinom(n = n, size = 1, prob = 0.5))
m_covariate_coefs <- log(0.9)
g_covariate_coefs <- log(1.1)
dat <- generate_full_data(m_fam = m_fam, m_intercept = m_intercept,
m_perturbation = m_perturbation, g_fam = g_fam, g_intercept = g_intercept,
g_perturbation = g_perturbation, pi = pi, n = n, B = 2,
covariate_matrix = covariate_matrix, m_covariate_coefs = m_covariate_coefs,
g_covariate_coefs = g_covariate_coefs, m_offset = m_offsets, g_offset = g_offsets)[[1]]
m <- dat$m
g <- dat$g
fit <- run_full_glmeiv_given_pilot_params(m = m, g = g, m_fam = m_fam, g_fam = g_fam,
pi_guess = 0.15, m_intercept_guess = log(0.1), m_perturbation_guess = log(1),
m_covariate_coefs_guess = log(1.4), g_intercept_guess = log(0.05),
g_perturbation_guess = log(1.4), g_covariate_coefs_guess = log(1.2),
covariate_matrix = covariate_matrix, m_offset = m_offsets, g_offset = g_offsets)
}
