% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_scale_em_algo_functs.R
\name{run_glmeiv_given_precomputations}
\alias{run_glmeiv_given_precomputations}
\title{Run GLM-EIV given precomputations}
\usage{
run_glmeiv_given_precomputations(
  m,
  g,
  m_precomp,
  g_precomp,
  covariate_matrix,
  m_offset,
  g_offset,
  n_em_rep,
  pi_guess_range,
  m_perturbation_guess_range,
  g_perturbation_guess_range,
  ep_tol = 1e-04
)
}
\arguments{
\item{m}{mRNA counts}

\item{g}{gRNA counts}

\item{m_precomp}{mRNA precomputation list, as outputted by \code{run_glmeiv_precomputation}}

\item{g_precomp}{gRNA precomputation list, as outputted by \code{run_glmeiv_precomputation}}

\item{covariate_matrix}{the matrix of covariates; NULL if there are no covariates.}

\item{m_offset}{offsets for m}

\item{g_offset}{offsets for g}

\item{n_em_rep}{number of EM reps to run on reduced GLM-EIV model}

\item{pi_guess_range}{range of values over which to sample pi}

\item{m_perturbation_guess_range}{range of values over which to sample m_perturbation}

\item{g_perturbation_guess_range}{range of values over which to sample g_perturbation}

\item{ep_tol}{(optional) EM convergence threshold}

\item{fam_str}{string describing family; currently "poisson" or "Negative Binomial" supported}
}
\value{
a fitted GLM-EIV model
}
\description{
Runs a full GLM-EIV given mRNA and gRNA precomputations, as outputted by \code{run_glmeiv_precomputation} function.
}
