% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_em_algo_functs.R
\name{run_glmeiv_known_p}
\alias{run_glmeiv_known_p}
\title{Run GLM-EIV with known p}
\usage{
run_glmeiv_known_p(
  m,
  g,
  m_fam,
  g_fam,
  covariate_matrix,
  p,
  m_offset = NULL,
  g_offset = NULL,
  n_runs = 4,
  p_flip = 0.15,
  ep_tol = 0.1,
  max_it = 50
)
}
\arguments{
\item{m}{m data vector}

\item{g}{g data vector}

\item{m_fam}{family object for m}

\item{g_fam}{family object for g}

\item{covariate_matrix}{the matrix of observed covariates}

\item{p}{the vector of p's; these typically are unobserved but will be used as the initial weights here}

\item{m_offset}{(default NULL) the vector of offsets for m}

\item{g_offset}{(default NULL) the vector of offsets for g}

\item{n_runs}{(default 10) number of EM algo runs}

\item{p_flip}{(default 0.15) expected fraction of initial weights to flip in each run}

\item{ep_tol}{(detault 0.1) tolerance threshold for EM convergence}

\item{max_it}{(default 50) maximum number of EM iterations (per run)}
}
\value{
the best EM run
}
\description{
Run GLM-EIV with known p
}
\examples{
dat <- get_quick_simulated_data(5000)
em_coefs <- run_glmeiv_known_p(m = dat$m, g = dat$g, m_fam = dat$m_fam, g_fam = dat$m_fam,
covariate_matrix = dat$covariate_matrix, p = dat$p, m_offset = NULL, g_offset = NULL)
# dat$m_coef; dat$g_coef; dat$pi
}
