% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_scale_em_algo_functs.R
\name{run_glmeiv_precomputation}
\alias{run_glmeiv_precomputation}
\title{Run GLM-EIV precomputation}
\usage{
run_glmeiv_precomputation(y, covariate_matrix, offset, fam)
}
\arguments{
\item{y}{a vector of gene or gRNA expressions}

\item{covariate_matrix}{the matrix of technical factors}

\item{offset}{a vector of offsets, generally log library size}

\item{fam_str}{a string specifying the family, currently either "Negative Binomial" or "poisson"}
}
\value{
a list containing (i) estimated theta (in NB case only), (ii) fitted values, (iii) fitted intercept term, (iv) covariate_coef terms
}
\description{
Runs a precomputation on a gene or gRNA.
}
