% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{run_glmeiv_random_init_simulatr}
\alias{run_glmeiv_random_init_simulatr}
\title{Run GLM-EIV (random inititalization)}
\usage{
run_glmeiv_random_init_simulatr(
  dat,
  m_fam,
  g_fam,
  covariate_matrix,
  m_offset,
  g_offset,
  alpha = 0.95,
  n_em_rep = 15,
  save_membership_probs_mult = 250,
  pi_guess_range = c(1e-05, 0.03),
  m_perturbation_guess_range = log(c(0.1, 1.5)),
  g_perturbation_guess_range = log(c(0.5, 10)),
  m_intercept_guess_range = log(c(1e-04, 0.1)),
  g_intercept_guess_range = log(c(1e-04, 0.1)),
  m_covariate_coefs_guess_range = log(c(0.5, 1.5)),
  g_covariate_coefs_guess_range = log(c(0.5, 1.5))
)
}
\arguments{
\item{m_fam}{family describing m}

\item{g_fam}{family describing g}

\item{covariate_matrix}{the matrix of covariates; NULL if there are no covariates.}

\item{m_offset}{offsets for m}

\item{g_offset}{offsets for g}

\item{alpha}{CIs returned at level (1-alpha)\\%}

\item{n_em_rep}{number of EM replicates}

\item{save_membership_probs_mult}{save membership probabilities at this multiple}

\item{pi_guess_range}{range of values over which to sample pi}

\item{m_perturbation_guess_range}{range of values over which to sample m_perturbation}

\item{g_perturbation_guess_range}{range of values over which to sample g_perturbation}

\item{m_intercept_guess_range}{range over which to sample m_intercept}

\item{g_intercept_guess_range}{range over which to sample g_intercept}

\item{m_covariate_coefs_guess_range}{range over which to sample m_covariate_coefs}

\item{g_covariate_coefs_guess_range}{range over which to sample g_covariate_coefs}
}
\value{

}
\description{
Runs GLM-EIV using random initializations for the parameters. Currently, the function assumes that there is at most a single covariate term.
}
