% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_matrix_functs.R
\name{run_inference_on_em_fit}
\alias{run_inference_on_em_fit}
\title{Run inference on EM fit}
\usage{
run_inference_on_em_fit(em_fit, alpha = 0.95)
}
\arguments{
\item{em_fit}{a fitted GLM-EIV model, as outputted by run_em_algo_given_init.}

\item{alpha}{(1-alpha) confidence interval produced}
}
\value{
a data frame with columns variable, estimate, std_error, p_value, confint_lower, confint_higher
}
\description{
Run inference on EM fit
}
\examples{
dat <- get_quick_simulated_data(1000)
em_fit <- run_em_algo_given_init(m = dat$m, g = dat$g, m_fam = poisson(), g_fam = poisson(),
covariate_matrix = dat$covariate_matrix, initial_Ti1s = dat$p, m_offset = NULL, g_offset = NULL)
se_table <- do_inference_on_em_fit(em_fit)
}
