% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding_algos.R
\name{run_thresholding_method}
\alias{run_thresholding_method}
\title{Run thresholding method}
\usage{
run_thresholding_method(
  m,
  m_fam,
  covariate_matrix,
  p_hat,
  m_offset = NULL,
  alpha = 0.95
)
}
\arguments{
\item{m}{a vector of m counts}

\item{m_fam}{a family object describing m}

\item{covariate_matrix}{the covariate matrix}

\item{p_hat}{the "estimated" vector of perturbations obtained through thresholding}

\item{m_offset}{(optional) offset term for m}

\item{alpha}{(optional; default .95) returns an alpha-CI for each estimated parameter}
}
\value{
a data frame of estimated coefficients, p-values, and standard errors
}
\description{
Runs the thresholding method.
}
\examples{
n <- 1000
m_fam <- poisson()
g_fam <- poisson()
m_offset <- g_offset <- NULL
pi <- 0.2
covariate_matrix <- NULL
m_coef <- c(2, -1)
g_coef <- c(-1, 2)
generated_data <- generate_data_from_model(m_fam, g_fam, m_coef, g_coef, pi, covariate_matrix, n = n)
m <- generated_data$m
g <- generated_data$g
p_hat <- threshold_counts_no_covariates(g_coef[1], g_coef[2], g, g_fam, pi)
run_thresholding_method(m, m_fam, covariate_matrix, p_hat)
}
