% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding_algos.R
\name{run_thresholding_method_simulatr}
\alias{run_thresholding_method_simulatr}
\title{Run thresholding method simulatr V2}
\usage{
run_thresholding_method_simulatr(
  dat,
  g_intercept,
  g_perturbation,
  g_fam,
  m_fam,
  pi,
  covariate_matrix,
  g_covariate_coefs,
  m_offset,
  g_offset,
  alpha
)
}
\arguments{
\item{g_fam}{family describing g}

\item{m_fam}{family describing m}

\item{covariate_matrix}{the matrix of covariates; NULL if there are no covariates.}

\item{m_offset}{offsets for m}

\item{g_offset}{offsets for g}
}
\value{
coefficient table
}
\description{
Runs the thresholding method
}
