% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_step_methods.R
\name{run_two_step_method_simulatr}
\alias{run_two_step_method_simulatr}
\title{Run two step method}
\usage{
run_two_step_method_simulatr(
  dat,
  g_intercept,
  g_perturbation,
  g_fam,
  m_fam,
  pi,
  covariate_matrix,
  g_covariate_coefs,
  m_offset,
  g_offset,
  alpha
)
}
\arguments{
\item{g_fam}{family describing g}

\item{m_fam}{family describing m}

\item{covariate_matrix}{the matrix of covariates; NULL if there are no covariates.}

\item{m_offset}{offsets for m}

\item{g_offset}{offsets for g}
}
\value{
coefficient table
}
\description{
Runs the method in which we (1) fit a mixture model to the grna modality, (2) assign perturbation identities based on posterior inclusion probabilities, and then (3) run the GLM
}
