% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_scale_em_algo_functs.R
\name{select_best_em_run}
\alias{select_best_em_run}
\title{Select best EM run}
\usage{
select_best_em_run(em_runs, m_perturbation_range = c(log(0.1), log(2)))
}
\arguments{
\item{em_runs}{a list of outputs of run_em_algo_given_init}

\item{m_perturbation_range}{range of acceptable values for m_perturbation}
}
\value{
the best run of the list
}
\description{
Returns the best run, defined as the run with m_perturbation in a reasonable range and
greatest log-likelihood.
}
