% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_initialization_functs.R
\name{threshold_counts}
\alias{threshold_counts}
\title{Threshold counts}
\usage{
threshold_counts(
  g,
  g_intercept,
  g_pert,
  g_fam,
  pi,
  covariate_matrix = NULL,
  g_covariate_coefs = NULL,
  g_offset = NULL
)
}
\arguments{
\item{g}{gRNA counts}

\item{g_intercept}{intercept for gRNA model}

\item{g_pert}{perturbation coefficient for gRNA model}

\item{g_fam}{family object describing distribution of g}

\item{pi}{probability of perturbation}

\item{covariate_matrix}{(optional) matrix of technical covariates}

\item{g_covariate_coefs}{(optional) coefficients corresponding to technical factors}

\item{g_offset}{(optional) the offset vector}
}
\value{

}
\description{
Thresholds counts according to the Bayes-optimal decision boundary. When a covariate matrix is present, the decision boundary that is used is the mean of the Bayes-optimal decision boundaries across all examples.
}
\examples{
g_intercept <- 2
g_pert <- 1
g_fam <- poisson()
pi <- 0.2
covariate_matrix <- NULL
g_covariate_coefs <- NULL
g_offset <- NULL
threshold_counts(g_intercept, g_pert, g_fam, pi, covariate_matrix, g_covariate_coefs, g_offset)
}
