% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_initialization_functs.R
\name{threshold_counts_no_covariates}
\alias{threshold_counts_no_covariates}
\title{Threshold counts (no covariates)}
\usage{
threshold_counts_no_covariates(g_intercept, g_pert, g, g_fam, pi)
}
\arguments{
\item{g_intercept}{(known) intercept of g model}

\item{g_pert}{(known) perturbation coefficient of g model}

\item{g}{g counts}

\item{g_fam}{family object describing g}

\item{pi}{(known) probability of perturbation}
}
\value{
The optimally thresholded counts p_hat
}
\description{
Threshold the g counts in the absence of covariates.
}
