% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_scale_em_algo_functs.R
\name{wrangle_glmeiv_result}
\alias{wrangle_glmeiv_result}
\title{Wrangle GLM-EIV result}
\usage{
wrangle_glmeiv_result(s, time, em_fit, exponentiate_coefs)
}
\arguments{
\item{s}{output of \code{run_inference_on_em_fit}}

\item{time}{execution time}

\item{em_fit}{a fitted GLM-EIV object}
}
\value{
data frame with results in tidy form
}
\description{
converts output of GLM-EIV into useful form.
}
