% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{create_simulatr_specifier_object}
\alias{create_simulatr_specifier_object}
\title{Create simulatr specifier object}
\usage{
create_simulatr_specifier_object(
  param_grid,
  fixed_params,
  one_rep_times,
  covariate_sampler = NULL
)
}
\arguments{
\item{param_grid}{a grid of parameters giving the parameter settings}

\item{fixed_params}{a list of fixed parameters}

\item{one_rep_times}{a named list giving the single rep time (either scalar or vector) of each method.}

\item{covariate_sampler}{a list of functions to generate a covariate matrix}
}
\value{
a simulatr_specifier object
}
\description{
Creates a simulatr_specifier object to run a simulation.
}
