% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{create_simulatr_specifier_object}
\alias{create_simulatr_specifier_object}
\title{Create simulatr specifier object (v2)}
\usage{
create_simulatr_specifier_object(
  param_grid,
  fixed_params,
  methods = c("glmeiv_slow", "glmeiv_fast", "thresholding", "unimodal_mixture")
)
}
\arguments{
\item{param_grid}{grid of parameters}

\item{fixed_params}{list of fixed parameters}

\item{methods}{a character vector giving the methods to run}

\item{one_rep_times}{vector of times for each method, as well as the data generation procedure}
}
\value{
a simulatr specifier object for use in simulatr
}
\description{
Create simulatr specifier object (v2)
}
