% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{create_simulatr_specifier_object}
\alias{create_simulatr_specifier_object}
\title{Create simulatr specifier object}
\usage{
create_simulatr_specifier_object(
  param_grid,
  fixed_params,
  covariate_sampler = NULL
)
}
\arguments{
\item{param_grid}{a grid of parameters giving the parameter settings}

\item{fixed_params}{a list of fixed parameters}

\item{covariate_sampler}{a list of functions to generate a covariate matrix}
}
\value{
a simulatr_specifier object
}
\description{
Creates a simulatr_specifier object to run a simulation.
}
\examples{
library(simulatr)
param_grid <- create_param_grid(varying_values = list(pi = seq(0.0, 0.5, 0.025),
m_perturbation = seq(0.0, -2, -0.1),
g_perturbation = seq(0.0, 2, 0.1)),
baseline_values = list(pi = 0.25, m_perturbation = -1, g_perturbation = 1))
fixed_params <- list(
 seed = 4,
 n = 1000,
 B = 2000,
 n_processors = 5,
 m_intercept = 2,
 g_intercept = 1,
 m_fam = poisson(),
 g_fam = poisson()
)

# simulation study 1: no covariates
covariate_sampler <- NULL

# simulation study 2: covariates present
covariate_sampler <- list(lib_size = function(n) rpois(n, 10),
                          p_mito = function(n) runif(n, 0, 10))
fixed_params[["m_covariate_coefs"]] <- c(0.2, -0.1)
fixed_params[["g_covariate_coefs"]] <- c(0.1, -0.2)
}
