% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr_functs.R
\name{generate_full_data}
\alias{generate_full_data}
\title{Generate full data}
\usage{
generate_full_data(
  m_fam,
  m_intercept,
  m_perturbation,
  g_fam,
  g_intercept,
  g_perturbation,
  pi,
  n,
  B,
  covariate_matrix,
  m_covariate_coefs,
  g_covariate_coefs,
  m_offset,
  g_offset
)
}
\arguments{
\item{m_fam}{family for m}

\item{m_intercept}{m intercept}

\item{m_perturbation}{m perturbation coefficient}

\item{g_fam}{family for g}

\item{g_intercept}{g intercept}

\item{g_perturbation}{g perturbation coefficient}

\item{pi}{probability of perturbation}

\item{n}{number of cells}

\item{B}{number of i.i.d. datasets to generate}

\item{covariate_matrix}{the (fixed) covariate matrix of confounding factors}

\item{m_covariate_coefs}{coefficients for technical factors in model for m}

\item{g_covariate_coefs}{coefficients for technical factors in model for g}

\item{m_offset}{optional (fixed) offset vector for m}

\item{g_offset}{optional (fixed) offset vector for g}
}
\value{
a list of length B of synthetic datasets with columns (p, m, g)
}
\description{
Generates B copies of a full GLM-EIV dataset.
}
\examples{
\dontrun{
library(magrittr)
m_fam <- g_fam <- poisson() \%>\% augment_family_object()
m_intercept <- 2; m_perturbation <- -1; g_intercept <- -2; g_perturbation <- 1
pi <- 0.2; n <- 1000; B <- 500
m_offset <- g_offset <- NULL
# no covariates
m_covariate_coefs <- g_covariate_coefs <- covariate_matrix <- NULL
dat_list <- generate_full_data(m_fam, m_intercept, m_perturbation, g_fam,
g_intercept, g_perturbation, pi, n, B, covariate_matrix, m_covariate_coefs,
g_covariate_coefs, m_offset, g_offset)
# with covariates
covariate_matrix <- data.frame(p_mito = runif(n, 0, 10))
m_covariate_coefs <- -0.1; g_covariate_coefs <- 0.2
dat_list <- generate_full_data(m_fam, m_intercept, m_perturbation, g_fam,
g_intercept, g_perturbation, pi, n, B, covariate_matrix, m_covariate_coefs,
g_covariate_coefs, m_offset, g_offset)
}
}
