% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory_functs.R
\name{get_tresholding_estimator_bias}
\alias{get_tresholding_estimator_bias}
\title{Get bias of thresholding estimator}
\usage{
get_tresholding_estimator_bias(m_perturbation, g_perturbation, pi)
}
\arguments{
\item{m_perturbation}{perturbation coefficient for mRNA model}

\item{g_perturbation}{perturbation coefficient for gRNA model}

\item{pi}{probability of perturbation}
}
\value{
the bias of the thresholding estimator
}
\description{
Assumes without loss of generality that m_intercept and g_intercept are 0. Also, assumes the Bayes-optimal threshold has been chosen.
}
