% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_em_algo_functs.R, R/core_em_algo_functs.R
\name{run_em_algo_multiple_inits}
\alias{run_em_algo_multiple_inits}
\alias{run_em_algo_given_init}
\title{Run EM algo given initialization weights}
\usage{
run_em_algo_multiple_inits(
  m,
  g,
  m_fam,
  g_fam,
  covariate_matrix,
  initial_Ti1_matrix,
  m_offset,
  g_offset,
  ep_tol = 0.5 * 1e-04,
  max_it = 50
)

run_em_algo_given_init(
  m,
  g,
  m_fam,
  g_fam,
  covariate_matrix,
  initial_Ti1s,
  m_offset,
  g_offset,
  ep_tol = 0.5 * 1e-04,
  max_it = 50
)
}
\arguments{
\item{m}{observed vector m}

\item{g}{observed vector g}

\item{m_fam}{augmented family of m}

\item{g_fam}{augmented family of g}

\item{covariate_matrix}{the matrix of covariates; set to NULL if there are no covariates.}

\item{initial_Ti1_matrix}{a matrix of starting membership probabilities}

\item{m_offset}{(optional) offsets for GLM for M}

\item{g_offset}{(optional) offsets for GLM for G}

\item{ep_tol}{EM convergence threshold}

\item{max_it}{maximum number of EM iterations}

\item{initial_Ti1s}{the initial vector of membership probabilities}
}
\value{
a list containing the following: fit object of M GLM, fit object of G GLM,
fit for pi, number of iterations,full log-likelihood of final model
}
\description{
Run EM algo given initialization weights
}
\examples{
n <- 1000
m_fam <- poisson()
g_fam <- poisson()
m_offset <- g_offset <- NULL
pi <- 0.2
covariate_matrix <- data.frame(p_mito = runif(n = n, 0, 10),
                               lib_size = rpois(n = n, lambda = 1))
m_coef <- c(-1, -2, 1, 0.5)
g_coef <- c(-1, 2, 1, 0.5)
generated_data <- generate_data_from_model(m_fam, g_fam, m_coef, g_coef, pi, covariate_matrix)
m <- generated_data$m
g <- generated_data$g
# run using single initialization
initial_Ti1s <- (g >= median(g))
em_res <- run_em_algo_given_init(m, g, m_fam, g_fam,
covariate_matrix, initial_Ti1s, m_offset, g_offset)

# run using multiple initializations
initial_Ti1_matrix <- replicate(n = 10, (initial_Ti1s + rbinom(n, 1, 0.1)) \%\% 2)
em_runs  <- run_em_algo_multiple_inits(m, g, m_fam, g_fam,
covariate_matrix,initial_Ti1_matrix, m_offset, g_offset)

# harder setting
n <- 5000
m_fam <- poisson()
g_fam <- poisson()
m_offset <- g_offset <- NULL
pi <- 0.2
covariate_matrix <- NULL
m_coef <- c(2, -2)
g_coef <- c(-2, 2)
generated_data <- generate_data_from_model(m_fam, g_fam, m_coef, g_coef,
pi, covariate_matrix, n = n)
m <- generated_data$m
g <- generated_data$g
# run using single initialization
initial_Ti1s <- generated_data$p
initial_Ti1_matrix <- replicate(n = 10, (initial_Ti1s + rbinom(n, 1, 0.3)) \%\% 2)
em_run <- run_em_algo_given_init(m, g, m_fam, g_fam, covariate_matrix, initial_Ti1s,
m_offset, g_offset)
em_runs  <- run_em_algo_multiple_inits(m, g, m_fam, g_fam,
covariate_matrix,initial_Ti1_matrix, m_offset, g_offset)
select_best_em_run(em_runs)
}
