% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding_algos.R
\name{run_thresholding_method}
\alias{run_thresholding_method}
\title{Run thresholding method}
\usage{
run_thresholding_method(
  phat,
  m,
  m_fam,
  m_offset,
  covariate_matrix,
  n_examples_per_param = 5,
  alpha = 0.95,
  exponentiate_coefs = TRUE
)
}
\arguments{
\item{phat}{the binary vector of thresholded gRNA counts}

\item{m}{mRNA counts}

\item{m_fam}{family object describing mRNA counts}

\item{m_offset}{offset for mRNA model}

\item{covariate_matrix}{matrix of cell-specific covariates}

\item{n_examples_per_param}{number of perturbed cells (per parameter) that must be present to attempt model fit}

\item{alpha}{returns alpha-level CIs.}
}
\value{
data frame of fitted parameters and CIs.
}
\description{
Runs the thresholding method on real data.
}
\details{
Note: consider combining this function with \code{run_thresholding_method_simulatr}.
}
