% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_em_algo_functs.R
\name{select_best_em_run}
\alias{select_best_em_run}
\title{Select best EM run}
\usage{
select_best_em_run(em_runs)
}
\arguments{
\item{em_runs}{a list of outputs of run_em_algo_given_init}
}
\value{
the best run of the list
}
\description{
Returns the best run, defined as the run with positive perturbation coefficient for G and
greatest log-likelihood.
}
